Fix MDX v1 remark/rehype plugin loading after migration
During the migration to MDX 1.0.0 (1) in GH-137 the now deprecated (2)
`mdPlugins` and `hastPlugins` options were not replaced with their
(new named) respective equivalents `remarkPlugins` and `rehypePlugins`.
Even if the documentation states that the options will be removed in v2
and are still supported (only showing a warning in the console when
still used), the defined plugins were not loaded anymore causing e.g.
no more automatic generation of `id` attributes for headers in MDX
content.

Therefore, to finish 100% of the migration, both options have been
renamed.

References:
  (1) https://mdxjs.com/blog/v1
  (2) https://mdxjs.com/blog/v1#deprecations

GH-144